API Classificação e Correção Model Ranking Global

+++ src/APIs/BibliotecaGamificada.Classificacao.Api/Models/RankingGlobal.cs
@@ -3,10 +3,10 @@ using BibliotecaGamificada.Instituicoes.Comum.Entidades;
 namespace BibliotecaGamificada.Classificacao.Api.Models
 {
     public class RankingGlobal{
-        public int quantidadePontos {get; set;}
+        public double quantidadePontos {get; set;}
         public Instituicao instituicao {get; set;}
 
-        public RankingGlobal(int quantidadePontos, Instituicao instituicao)
+        public RankingGlobal(double quantidadePontos, Instituicao instituicao)
         {
             this.quantidadePontos = quantidadePontos;
             this.instituicao = instituicao;
